requirements.txt
Flask
requests
google-cloud-storage
opencv-python
numpy
imageio-ffmpeg
twilio
python-dotenv
Flask-CORS
ffmpeg